+++ geode-core/src/main/java/org/apache/geode/cache/client/internal/ClientMetadataService.java
@@ -539,7 +539,7 @@ public class ClientMetadataService {
           }
         }
       };
-      SingleHopClientExecutor.submitTask(fetchTask);
+      SingleHopClientExecutorImpl.getInstance().submitTask(fetchTask);
     }
   }
 
@@ -652,7 +652,7 @@ public class ClientMetadataService {
           }
         }
       };
-      SingleHopClientExecutor.submitTask(fetchTask);
+      SingleHopClientExecutorImpl.getInstance().submitTask(fetchTask);
     }
   }
