+++ geode-core/src/main/java/org/apache/geode/cache/client/internal/PutAllOp.java
@@ -111,8 +111,9 @@ public class PutAllOp {
         new HashMap<ServerLocation, RuntimeException>();
     PutAllPartialResult result = new PutAllPartialResult(map.size());
     try {
-      Map<ServerLocation, Object> results = SingleHopClientExecutor.submitBulkOp(callableTasks, cms,
-          (LocalRegion) region, failedServers);
+      Map<ServerLocation, Object> results = SingleHopClientExecutorImpl.getInstance()
+          .submitBulkOp(callableTasks, cms,
+              (LocalRegion) region, failedServers);
       for (Map.Entry<ServerLocation, Object> entry : results.entrySet()) {
         Object value = entry.getValue();
         if (value instanceof PutAllPartialResultException) {
